requirements.txt
Flask
openai>=1.0.0
networkx
pydot
python-dotenv
gunicorn
graphviz
